+++ API/src/Modules/XpSystems/Momentum.XpSystems.Application/DTOs/Cosmetic/LevelsDto.cs
@@ -6,10 +6,10 @@
         public int StartingValue { get; set; }
         public int LinearScaleBaseIncrease { get; set; }
         public int LinearScaleInterval { get; set; }
-        public double LinearScaleIntervalMultiplier { get; set; }
+        public decimal LinearScaleIntervalMultiplier { get; set; }
         public int StaticScaleStart { get; set; }
-        public double StaticScaleBaseMultiplier { get; set; }
+        public decimal StaticScaleBaseMultiplier { get; set; }
         public int StaticScaleInterval { get; set; }
-        public double StaticScaleIntervalMultiplier { get; set; }
+        public decimal StaticScaleIntervalMultiplier { get; set; }
     }
 }
