+++ API/src/Modules/XpSystems/Momentum.XpSystems.Application/DTOs/Rank/GroupsDto.cs
@@ -3,9 +3,9 @@
     public class GroupsDto
     {
         public int MaxGroups { get; set; }
-        public double[] GroupScaleFactors { get; set; }
-        public double[] GroupExponents { get; set; }
+        public decimal[] GroupScaleFactors { get; set; }
+        public decimal[] GroupExponents { get; set; }
         public int[] GroupMinimumSizes { get; set; }
-        public double[] GroupPointPercentages { get; set; }
+        public decimal[] GroupPointPercentages { get; set; }
     }
 }
